+++ aws-sdk-core/lib/aws-sdk-core.rb
@@ -205,8 +205,9 @@ module Aws
   # @api private
   module Stubbing
     autoload :EmptyStub, 'aws-sdk-core/stubbing/empty_stub'
-    autoload :StubData, 'aws-sdk-core/stubbing/stub_data'
     autoload :DataApplicator, 'aws-sdk-core/stubbing/data_applicator'
+    autoload :StubData, 'aws-sdk-core/stubbing/stub_data'
+    autoload :XmlError, 'aws-sdk-core/stubbing/xml_error'
     module Protocols
       autoload :EC2, 'aws-sdk-core/stubbing/protocols/ec2'
       autoload :Json, 'aws-sdk-core/stubbing/protocols/json'