+++ aws-sdk-core/spec/aws/stubbing/protocols/rest_xml_spec.rb
@@ -93,6 +93,18 @@ module Aws
             JSON
           end
 
+          it 'can stub errors' do
+            resp = RestXml.new.stub_error('error-code')
+            expect(normalize(resp.body.string)).to eq(normalize(<<-XML))
+              <ErrorResponse>
+                <Error>
+                  <Code>error-code</Code>
+                  <Message>stubbed-response-error-message</Message>
+                </Error>
+              </ErrorResponse>
+            XML
+          end
+
         end
       end
     end